translation strings and cleanup
